+++ CHANGELOG.md
@@ -19,6 +19,7 @@ however, insignificant breaking changes do not guarantee a major version bump, s
 - Added `confirm_thread_creation`, `confirm_thread_creation_title`, `confirm_thread_creation_description`, `confirm_thread_creation_accept`, `confirm_thread_creation_deny` to allow users to confirm that they indeed want to create a new thread. ([GH #2773](https://github.com/kyb3r/modmail/issues/2773))
 - Support Gyazo image links in message embeds. ([GH #282](https://github.com/kyb3r/modmail/issues/282))
 - Added `silent` argument to `?contact` to restore old behaviour.
+- If `?help` is sent, bot does checks on every command, `?help all` restores old behaviour. ([GH #2847](https://github.com/kyb3r/modmail/issues/2847))
 
 ### Fixed